requirements.txt
aiohttp==2.2.5
async-timeout==1.3.0
chardet==3.0.4
multidict==3.1.3
SQLAlchemy==1.1.14
transaction==2.1.2
yarl==0.12.0
zope.interface==4.4.2
zope.sqlalchemy==0.7.7
